fix genus representatives and tidy up
