+++ src/QuadForm/Herm/Genus.jl
@@ -1,7 +1,6 @@
 export genus, representative, rank, det, uniformizer, det_representative,
-       gram_matrix, representative, genus, genera_hermitian,
-       local_genera_hermitian, rank, orthogonal_sum, isinert, scales, ranks,
-       dets, issplit
+       gram_matrix, genera_hermitian, local_genera_hermitian, rank, 
+       orthogonal_sum, isinert, scales, ranks, dets, issplit
 
 ################################################################################
 #
@@ -1014,6 +1013,55 @@ end
 
 ################################################################################
 #
+#  Sum of local genera
+#
+################################################################################
+
+function orthogonal_sum(G1::LocalGenusHerm, G2::LocalGenusHerm)
+  @req prime(G1) == prime(G2) "Local genera must have the same prime ideal"
+  if !G1.isdyadic || !G2.isramified
+    return _direct_sum_easy(G1, G2)
+  else
+    L1 = representative(G1)
+    L2 = representative(G2)
+    L3, = orthogonal_sum(L1, L2)
+    return genus(L3, prime(G1))
+  end
+end
+
+function _direct_sum_easy(G1::LocalGenusHerm, G2::LocalGenusHerm)
+  # We do a merge sort
+  i1 = 1
+  i2 = 1
+  z = Tuple{Int, Int, Int}[]
+  while i1 <= length(G1) && i2 <= length(G2)
+    if scale(G1, i1) < scale(G2, i2)
+      push!(z, G1.data[i1])
+      i1 += 1
+    elseif scale(G2, i2) < scale(G1, i1)
+      push!(z, G2.data[i2])
+      i2 += 1
+    else
+      @assert scale(G1, i1) == scale(G2, i2)
+      push!(z, (scale(G1, i1), rank(G1, i1) + rank(G2, i2), det(G1, i1) * det(G2, i2)))
+      i1 += 1
+      i2 += 1
+    end
+  end
+
+  if i1 <= length(G1)
+    append!(z, G1.data[i1:length(G1)])
+  end
+
+  if i2 <= length(G2)
+    append!(z, G2.data[i2:length(G2)])
+  end
+
+  return genus(HermLat, base_field(G1), prime(G1), z)
+end
+
+################################################################################
+#
 #  Global genus
 #
 ################################################################################
@@ -1627,312 +1675,3 @@ function genera_hermitian(E, rank, signatures, determinant; max_scale = nothing)
   return res
 end
 
-################################################################################
-#
-#  Genus representatives
-#
-################################################################################
-
-
-@doc Markdown.doc"""
-    smallest_neighbour_prime(L::HermLat) -> Bool, NfRelOrdIdl, Vector{NfOrdIdl}
-
-Given a hermitian lattice `L`, return `def, P0, bad` such that:
-
-- `def` is `true` if `L` is definite, else `false`;
-- `P0` is a prime ideal in the base ring `O` of `L` which is not bad, such that
-  `L` is isotropic at `minimum(P0)` and `P0` has smallest minimum among the primes 
-  satisfying these properties; if `L` is indefinite, `P0` is set to be the trivial ideal
-- `bad` is a vector of prime ideals `p` in the maximal order of the fixed field 
-  of `L` such that `L_p` is not modular or `p` is dyadic and is not coprime to
-  the discriminant of `O`.
-"""
-function smallest_neighbour_prime(L)
-  S = base_ring(L)
-  R = base_ring(S)
-  lp = bad_primes(L)
-  bad = ideal_type(R)[p for p in lp if !ismodular(L, p)[1] ]
-  for (p,_) in factor(discriminant(S))
-    if isdyadic(p) && !(p in bad)
-      push!(bad, p)
-    end
-  end
-
-  if !isdefinite(L)
-    return false, 1*S, bad
-  end
-
-  # TODO: This does not find the prime ideal with smallest norm,
-  # but with smallest minimum ...
-
-  m = rank(L)
-  p = 1
-  P = ideal_type(R)[]
-  while length(P) == 0
-    p = next_prime(p)
-    lp = ideal_type(R)[ p[1] for p in prime_decomposition(R, p)]
-    P = setdiff(lp, bad)
-    if m == 2
-      P = filter(p -> isisotropic(L, p), P)
-    end
-  end
-  Q = prime_decomposition(S, P[1])[1][1]
-
-  if isempty(bad)
-    I = 1 * S
-  else
-    I = prod(bad) * S
-  end
-  n = absolute_norm(Q)
-  if n >= 1000
-    PP = prime_ideals_up_to(S, 1000)
-    for QQ in PP
-      if !iscoprime(QQ, I)
-        continue
-      end
-
-      if isisotropic(L, QQ)
-        return true, QQ, bad
-      end
-    end
-  end
-  PP = prime_ideals_up_to(S, n)
-  for QQ in PP
-    if !iscoprime(QQ, I)
-      continue
-    end
-    if isisotropic(L, QQ)
-      return true, QQ, bad
-    end
-  end
-  throw(error("Impossible"))
-end
-
-@doc Markdown.doc"""
-    genus_generators(L::HermLat) -> Vector{Tuple{NfRelOrdIdl, fmpz}}, Bool, 
-                                    NfRelOrdIdl
-
-Given a hermitian lattice `L`, return `gens, def, P0` such that:
-
-- `gens` is a vector of tuples `(P,e)` consisting of a prime ideal `P` in the base ring of `L`
-  and an integer `e \geq 2` which can be used to compute the ideal `\mathfrak A` in line 11
-  of [Kir19, Algorithm 4.7.]); 
-- `def` is `true` if `L` is definite, else `false`;
-- `P0` is a prime ideal in the base ring of `L` which is not bad, such that
-  `L` is isotropic at `minimum(P0)` and `P0` has smallest minimum among the primes 
-  satisfying these properties.
-"""
-function genus_generators(L::HermLat)
-  R = base_ring(L)
-  E = nf(R)
-  D = different(R)
-  a = involution(L)
-  def, P0, bad = smallest_neighbour_prime(L)
-
-  local bad_prod::ideal_type(base_ring(R))
-
-  if isempty(bad)
-    bad_prod = 1 * base_ring(R)
-  else
-    bad_prod = prod(bad)
-  end
-
-  # First the ideals coming from the C/C0 quotient
-  Eabs, EabstoE = absolute_simple_field(ambient_space(L))
-  Rabs = maximal_order(Eabs)
-  C, h = class_group(Rabs)
-  RR = base_ring(R)
-  C0 = support(D)
-  CC, hh = class_group(RR)
-  for p in find_gens(pseudo_inv(hh), PrimesSet(2, -1))[1]
-    if !(p in C0)
-      push!(C0, sum(R * R(EabstoE(elem_in_nf(b))) for b in basis(p)))
-    end
-  end
-  Q0, q0 = quo(C, elem_type(C)[ h\ideal(Rabs, [Rabs(EabstoE\b) for b in absolute_basis(i)]) for i in C0])
-  q00 = pseudo_inv(q0) * h
-  PP = ideal_type(R)[]
-
-  local F::GaloisField
-
-  local W::Generic.QuotientModule{gfp_elem}
-
-  if iseven(rank(L))
-    for (P, e) in factor(D)
-      p = minimum(P)
-      G = genus(L, p)
-      if any(i -> isodd(rank(G, i)), 1:length(G))
-        continue
-      elseif !isdyadic(p)
-        if any(i -> iseven(scale(G, i)), 1:length(G))
-          continue
-        end
-      else
-        if any(i -> isodd(e  + scale(G, i)) || (e + scale(G, i) != G.ni[i]), 1:length(G))
-          continue
-        end
-      end
-      push!(PP, P)
-    end
-    
-    if !isempty(PP)
-      U, f = unit_group_fac_elem(Rabs)
-      UU, ff = unit_group_fac_elem(RR)
-      nnorm = hom(U, UU, GrpAbFinGenElem[ff\FacElem(nf(RR)(norm(f(U[i])))) for i in 1:ngens(U)])
-      l = length(PP)
-      VD = Int[ valuation(D, P) for P in PP ]
-      K, k = kernel(nnorm)
-      F = GF(2, cached = false)
-      V = VectorSpace(F, length(PP))
-      S = elem_type(V)[]
-      for u in gens(K)
-        z = elem_type(F)[]
-        for i in 1:length(PP)
-          zz = R(EabstoE(evaluate(f(k(u))))) - 1
-          if iszero(zz) || (valuation(zz, PP[i]) >= VD[i])
-            push!(z, F(0))
-          else
-            push!(z, F(1))
-          end
-        end
-        push!(S, V(z)::elem_type(V))
-      end
-      _T, _ = sub(V, S)
-      W, w = quo(V, _T)
-      if dim(W) == 0
-        PP = ideal_type(R)[]
-      end
-    end
-  end
-
-  Gens = Tuple{ideal_type(R), fmpz}[]
-
-  if isempty(PP)
-    S = GrpAbFinGenElem[]
-    Q, q = quo(Q0, S)::Tuple{GrpAbFinGen, GrpAbFinGenMap}
-    Work = def ? typeof(P0)[ P0 ] : typeof(P0)[]
-    p = 2
-    while order(Q) > 1
-      while isempty(Work)
-        p = next_prime(p)
-        Work = ideal_type(R)[ QQ for QQ in support(p * R) if issplit(QQ) && valuation(bad, QQ) == 0 ]
-      end
-      P = popfirst!(Work)
-      c = (q00\P)::GrpAbFinGenElem
-      o = order(q(c))::fmpz
-      if !isone(o)
-        push!(S, c)
-        Q, q = quo(Q0, S)::Tuple{GrpAbFinGen, GrpAbFinGenMap}
-        push!(Gens, (P, o))
-      end
-    end
-  else
-    ll = Int(order(Q0))
-    cocycle = Matrix{elem_type(W)}(undef, ll, ll)
-    for i in 1:ll
-      for j in 1:ll
-        cocycle[i,j] = zero(W)
-      end
-    end
-    C = collect(Q0)
-    ideals = ideal_type(Rabs)[ q00(C[i]) for i in 1:length(C) ]
-    for i in 1:ll
-      for j in 1:i
-        ij = findfirst(isequal(C[i] + C[j]), C)
-        Iabs = ideals[i] * ideals[j] * inv(ideals[ij])
-        I = EabstoE(Iabs)
-        J = I * inv(a(I))
-        Jabs = EabstoE\J
-        ok, x = isprincipal(Jabs)
-        u = f(nnorm\(-(ff\FacElem(nf(RR)(norm(x))))))
-        x = x * u
-        @assert norm(x) == 1
-        if evaluate(x) == 1
-          y = w(V(zeros(F,length(PP))))
-        else
-          y = w(V([ valuation(evaluate(x) - 1, PP[i]) >= VD[i] ? F(0) : F(1) for i in 1:length(PP)]))
-        end
-        cocycle[i, j] = y
-        cocycle[j, i] = y
-      end
-    end
-
-    S = Tuple{elem_type(Q0), Generic.QuotientModuleElem{elem_type(F)}}[(id(Q0), zero(W))]
-    Work = def ? typeof(P0)[ P0 ] : typeof(P0)[]
-    p = 2
-    while length(S) != order(Q0) * dim(W)
-      while isempty(Work)
-        p = next_prime(p)
-        Work = ideal_type(R)[ QQ for QQ in support(p * R) if issplit(QQ) && valuation(bad, QQ) == 0 ]
-      end
-      P = popfirst!(Work)
-      c = q00\P
-      i = findfirst(isequal(c), C)
-      Iabs = P * inv(ideals[i])
-      I = EabstoE(Iabs)
-      J = I * inv(a(I))
-      Jabs = EabstoE\J
-      ok, x = isprincipal(Jabs)
-      u = f(nnorm\(-(ff\FacElem(nf(RR)(norm(x))))))
-      x = x * u
-      @assert norm(x) == 1
-      if evaluate(x) == 1
-        y = w(V(zeros(F,length(PP))))
-      else
-        y = w(V([ valuation(evaluate(x) - 1, PP[i]) >= VD[i] ? F(0) : F(1) for i in 1:length(PP)]))
-      end
-      idx = findfirst(isequal(P), PP)
-      if idx !== nothing
-        y = V(elem_type(F)[i == idx ? y[i] + 1 : y[i]  for i in 1:dim(V)])
-      end
-      elt = (c, w(y))
-      elt1 = elt
-      o = 1
-      siz = length(S)
-      while !(elt1 in S)
-        j = findfirst(isequal(elt1[1]), C)
-        @assert !isnothing(j)
-        for l in 1:siz
-          elt2 = S[l]
-          k = findfirst(isequal(elt2[1]), C)
-          @assert !isnothing(k)
-          prod = (elt1[1] * elt2[1], elt1[2] + elt2[2] + cycycle[j, k])
-          if !(prod in S)
-            push!(S, prod)
-          end
-        end
-        elt1 = (elt[1] * elt1[1], elt[2] + elt[1] + cocycle[i, j])
-        o = o + 1
-      end
-      @assert length(S) == siz * o
-      if o != 1
-        push!(Gens, (P, o))
-      end
-    end
-  end
-
-  return Gens, def, P0
-end
-
-function representatives(G::GenusHerm)
-  return genus_representatives(representative(G))
-end
-
-function _all_row_span(M)
-  F = base_ring(M)
-  rows = Vector{Vector{elem_type(F)}}(undef, nrows(M))
-  for i in 1:nrows(M)
-    rows[i] = elem_type(F)[M[i, j] for j in 1:ncols(M)]
-  end
-  n = nrows(M)
-  it = Iterators.product([F for i in 1:n]...)
-  res = Vector{Vector{elem_type(F)}}()
-  for c in it
-    z = Ref(c[1]) .* rows[1]
-    for i in 2:n
-      z = z .+ (Ref(c[i]) .* rows[i])
-    end
-    push!(res, z)
-  end
-  return res
-end
