+++ test/QuadForm/Herm.jl
@@ -1,4 +1,5 @@
 @testset "Herm" begin
   include("Herm/Spaces.jl")
   include("Herm/Genus.jl")
+  include("Herm/GenusRep.jl")
 end
